move increase batch size method
